skale-analytics-load/requirements.txt
skale.py==4.1b22
python-dotenv==0.14.0
requests~=2.23.0
web3~=5.6.0
mysql-connector-python~=8.0.21
